+++ demisto_sdk/commands/format/update_generic.py
@@ -1,21 +1,23 @@
 import os
 import re
 from copy import deepcopy
-from typing import Set, Union
+from typing import Optional, Set, Union
 
 import click
 import yaml
-from demisto_sdk.commands.common.constants import FileType
+from demisto_sdk.commands.common.constants import INTEGRATION, FileType
 from demisto_sdk.commands.common.hook_validations.structure import \
     StructureValidator
 from demisto_sdk.commands.common.tools import (LOG_COLORS, find_type,
                                                get_dict_from_file,
+                                               get_pack_metadata,
                                                get_remote_file,
                                                is_file_from_content_repo,
                                                print_color)
 from demisto_sdk.commands.format.format_constants import (
     DEFAULT_VERSION, ERROR_RETURN_CODE, NEW_FILE_DEFAULT_5_FROMVERSION,
-    OLD_FILE_DEFAULT_1_FROMVERSION, SKIP_RETURN_CODE, SUCCESS_RETURN_CODE)
+    OLD_FILE_DEFAULT_1_FROMVERSION, SKIP_RETURN_CODE, SUCCESS_RETURN_CODE,
+    VERSION_6_0_0)
 from ruamel.yaml import YAML
 
 ryaml = YAML()
@@ -204,29 +206,39 @@ class BaseUpdate:
                 return reg
         return None
 
-    def set_fromVersion(self, from_version=None):
+    def set_fromVersion(self, from_version=None, file_type: Optional[str] = None):
         """Sets fromversion key in file:
         Args:
             from_version: The specific from_version value.
+            file_type: what is the file type: for now only integration type passed
         """
+        metadata = get_pack_metadata(self.source_file)
+        # if it is new contributed pack = setting version to 6.0.0
+        should_set_from_version = ((metadata.get('currentVersion', '') == '1.0.0') and (metadata.get('support', '') != 'xsoar'))
+
         # If there is no existing file in content repo
         if not self.old_file:
             if self.verbose:
                 click.echo('Setting fromVersion field')
             # If current file does not have fromversion key
             if self.from_version_key not in self.data:
-
                 # If user entered specific from version key to be set
                 if from_version:
                     self.data[self.from_version_key] = from_version
-
+                # if it is new contributed pack = setting version to 6.0.0
+                elif should_set_from_version:
+                    self.data[self.from_version_key] = VERSION_6_0_0
                 # Otherwise add fromversion key to current file and set to default 5.0.0
                 else:
                     self.data[self.from_version_key] = NEW_FILE_DEFAULT_5_FROMVERSION
-
             # If user wants to modify fromversion key and the key already existed
             elif from_version:
                 self.data[self.from_version_key] = from_version
+            # if it is new contributed pack, this is integration, and its version is 5.5.0 do not change it
+            # if it is new contributed pack = setting version to 6.0.0
+            elif should_set_from_version:
+                if self.data.get(self.from_version_key) != '5.5.0' or file_type != INTEGRATION:
+                    self.data[self.from_version_key] = VERSION_6_0_0
 
         # If there is an existing file in content repo
         else: