+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 * Added a validation that core packs are not dependent on non-core packs.
 * Fixed an issue where in some cases the `get_remote_file` function failed due to an invalid path.
 * Fixed an issue where the **create-id-set** command did not identify unified integrations correctly.
-* When formatting new contributors packs, the **format** command will set *fromversion* to 6.0.0 unless stated another version.
+* When formatting new contributed pack, the **format** command will set the *fromversion* key to 6.0.0 in the relevant files.
 
 # 1.3.5
 * Added a validation that layoutscontainer's id and name are matching. Updated the format of layoutcontainer to include update_id too.