+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 * Added a validation that core packs are not dependent on non-core packs.
+* Added a validation that a pack name follows XSOAR standards.
 * Fixed an issue where in some cases the `get_remote_file` function failed due to an invalid path.
 * Fixed an issue where running **update-release-notes** with updated integration logo, did not detect any file changes.
 * Fixed an issue where the **create-id-set** command did not identify unified integrations correctly.